Created using Colab
This is a Project Test for a Supervised Learning CNN. 
The result was not necessarily the concern, but rather, the journey of trying different data augmentations, layer units, metrics and beyond. I'm placing this under the K_Means_Events, though not explicitly related, because it's still a project I hope can help others on their machine learning journey.